// src/components/crm/client-details/types/MonthlyForecast.ts
export interface MonthlyForecast {
  month: string;
  amount: number;
}

export interface MonthlyForecastColumns {
  forecast_jan: number;
  forecast_feb: number;
  forecast_mar: number;
  forecast_apr: number;
  forecast_may: number;
  forecast_jun: number;
  forecast_jul: number;
  forecast_aug: number;
  forecast_sep: number;
  forecast_oct: number;
  forecast_nov: number;
  forecast_dec: number;
}